Added more tests. Still investigating delete cards issue.

+++ src/main/java/com/hercule/eshop/services/UserServiceImpl.java
@@ -4,6 +4,7 @@ import com.hercule.eshop.models.Cart;
 import com.hercule.eshop.models.User;
 import com.hercule.eshop.repositories.RoleRepository;
 import com.hercule.eshop.repositories.UserRepository;
+import com.hercule.eshop.stripe.models.StripeCustomer;
 import com.hercule.eshop.stripe.services.StripeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -60,7 +61,12 @@ public class UserServiceImpl implements UserService
         Cart cart = cartService.findCartByUserId(user);
         cartService.deleteCart(cart);
 
-        stripeService.deleteCustomer(user);
+        StripeCustomer stripeCustomer = stripeService.getCustomerByUserId(user.getId());
+        if (stripeCustomer != null)
+        {
+            stripeService.deleteCustomer(user);
+        }
+
         userRepository.delete(user);
     }
